project_mercury/rtm/src/main/java/com/smartsparrow/rtm/subscription/courseware/created/ComponentCreatedRTMConsumable.java
package com.smartsparrow.rtm.subscription.courseware.created;

import com.smartsparrow.rtm.message.RTMClientContext;
import com.smartsparrow.rtm.subscription.courseware.ActivityRTMSubscription;
import com.smartsparrow.rtm.subscription.courseware.message.ComponentCreatedBroadcastMessage;
import com.smartsparrow.rtm.subscription.data.AbstractRTMConsumable;
import com.smartsparrow.pubsub.data.RTMEvent;

/**
 * This RTM consumable describes a component created event
 */
public class ComponentCreatedRTMConsumable extends AbstractRTMConsumable<ComponentCreatedBroadcastMessage> {

    private static final long serialVersionUID = 7878795094033088633L;

    public ComponentCreatedRTMConsumable(RTMClientContext rtmClientContext, ComponentCreatedBroadcastMessage content) {
        super(rtmClientContext, content);
    }

    @Override
    public RTMClientContext getRTMClientContext() {
        return rtmClientContext;
    }

    @Override
    public String getName() {
        return String.format("author.activity/%s/%s", content.getActivityId(), getRTMEvent().getName());
    }

    @Override
    public String getSubscriptionName() {
        return ActivityRTMSubscription.NAME(content.getActivityId());
    }

    @Override
    public RTMEvent getRTMEvent() {
        return new ComponentCreatedRTMEvent();
    }

    @Override
    public boolean equals(Object o) {
        return super.equals(o);
    }

    @Override
    public int hashCode() {
        return super.hashCode();
    }
}
